% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bound_BE.R
\name{Bound_BE}
\alias{Bound_BE}
\title{Produce a Berry-Esseen-type bound}
\usage{
Bound_BE(
  setup = list(continuity = FALSE, iid = FALSE, no_skewness = FALSE),
  n,
  K4,
  K3 = NULL,
  lambda3 = NULL,
  K3tilde = NULL,
  regularity = list(C0 = 1, p = 2, kappa = 0.99),
  eps = 0.1
)
}
\arguments{
\item{setup}{logical vector of size 3 made up of
the following components: \itemize{
   \item \code{continuity}: if \code{TRUE}, assume that the distribution is continuous.

   \item \code{iid}: if \code{TRUE}, assume that the random variables are i.i.d.

   \item \code{no_skewness}: if \code{TRUE}, assume that the distribution is unskewed.
}}

\item{regularity}{list of length up to three
(only used in the `continuity=TRUE` framework)
with the following components:\itemize{

   \item `C0` and `p`: only used in the `iid=FALSE` case.
   It corresponds to the assumption of a polynomial bound on f_Sn:
   \eqn{|f_Sn(u)| <= C_0 * u^(-p)} for every \eqn{u > a_n},
   where \eqn{a_n := 2t_1^* \pisqrt(n) / K3tilde}.

   \item `kappa`: only used in the `iid=TRUE` case.
   Corresponds to a bound on the modulus of the characteristic function of
   the standardized \eqn{X_n}. More precisely, `kappa` is an upper bound on

   kappa = sup of modulus of f_{X_n / sigma_n}(t)
   over all t such that \eqn{|t| >= 2 t_1^* \pi / K3tilde}
}}
}
\description{
This bounds follows from the triangular inequality
and the bound on the difference
between a cdf and its 1st-order Edgeworth Expansion
}
\examples{
setup = list(continuity = FALSE, iid = FALSE, no_skewness = FALSE)
regularity = list(C0 = 1, p = 2, kappa = 0.99)

#computedBound_EE1 <- Bound_EE1(
#  setup = setup, n = 150, K4 = 9,
#  regularity = regularity, eps = 0.1 )

#computedBound_BE <- Bound_BE(
#  setup = setup, n = 150, K4 = 9,
#  regularity = regularity, eps = 0.1 )

#print(c(computedBound_EE1, computedBound_BE))

}
