% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_power.R
\name{t_test_powerAnalysis}
\alias{t_test_powerAnalysis}
\title{Computation of power and sufficient sample size for the t-test}
\usage{
t_test_powerAnalysis(
  eta = NULL,
  n = NULL,
  beta = NULL,
  alpha = 0.05,
  K4 = 9,
  kappa = 0.99
)
}
\arguments{
\item{eta}{the effect size \mjseqn{\eta} that
characterizes the alternative hypothesis}

\item{n}{sample size}

\item{beta}{the power of detecting the effect \code{eta} using the sample size \code{n}}

\item{alpha}{the level of the test}

\item{K4}{the kurtosis of the \mjseqn{X_i}}

\item{kappa}{Regularity parameter of the distribution of the \mjseqn{X_i}
It corresponds to a bound on the modulus of the characteristic function
of the standardized \mjseqn{X_n}.
More precisely, \code{kappa} is an upper bound on
\mjseqn{kappa :=} sup of modulus of \mjseqn{f_{X_n / \sigma_n}(t)}
over all \mjseqn{t} such that \mjseqn{|t| \geq 2 t_1^* \pi / K3tilde}.}
}
\description{
Let \mjseqn{X_1, \dots, X_n} be \mjseqn{n} i.i.d. variables
with mean \mjseqn{\mu}, variance \mjseqn{\sigma^2}.
Assume that we want to test the hypothesis
\mjseqn{H_0: \mu \leq \mu_0} against the alternative \mjseqn{H_1: \mu \leq \mu_0}.
Let \mjseqn{\eta := (\mu - \mu_0) / \sigma} be the effect size,
i.e. the distance between the null and the alternative hypotheses,
measured in terms of standard deviations.
There is a relation between the sample size \code{n}, the effect size \code{eta}
and the power \code{beta}.
}
\details{
\loadmathjax
}
\examples{

# Sufficient sample size to detect an effect of 0.5 standard deviation with probability 80\%
t_test_powerAnalysis(eta = 0.5, beta = 0.8)
# We can detect an effect of 0.5 standard deviations with probability 80\% for n >= 548

# Power of an experiment to detect an effect of 0.5 with a sample size of n = 800
t_test_powerAnalysis(eta = 0.5, n = 800)
# We can detect an effect of 0.5 standard deviations with probability 85.1\% for n = 800

# Smallest effect size that can be detected with a probability of 80\% for a sample size of n = 800
t_test_powerAnalysis(n = 800, beta = 0.8)
# We can detect an effect of 0.114 standard deviations with probability 80\% for n = 800


}
\references{
Derumigny A., Girard L., and Guyonvarch Y. (2021).
Explicit non-asymptotic bounds for the distance to the first-order Edgeworth expansion,
ArXiv preprint \href{https://arxiv.org/abs/2101.05780}{arxiv:2101.05780}.
}
